Do not paginate if limit param is present
Kaminari uses limit() internally to paginate. If limit is also set
externally, it can either mess with the pagination or with the
limiting itself (depending on which one is done first). Here I'm
forcing the pagination to only take place if no limit is passed as
param; otherwise, the limit takes priority.

This way, both pageable and limitable can be set in any order and
the behaviour stays the same. The same logic applies on limit.rb.